requirements.txt

aiohttp==3.9.1
websockets==12.0
numpy==1.26.2
pandas==2.1.4
python-dotenv==1.0.0
# bitfinex-api-py # Optional if we used the wrapper, but implemented custom clients.
